* fixed bug introduced in revision 3335: Hibyte lost in channel mask when restoring state from XML
git-svn-id: svn://localhost/ardour2/branches/3.0@4356 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/midi_channel_selector.cc
@@ -160,7 +160,7 @@ MidiMultipleChannelSelector::~MidiMultipleChannelSelector()
 }
 
 void
-MidiMultipleChannelSelector::set_channel_mode(ChannelMode mode, uint8_t mask)
+MidiMultipleChannelSelector::set_channel_mode(ChannelMode mode, uint16_t mask)
 {
 	switch (mode) {
 	case AllChannels:
